fix—buddy_release_block() logs, sets errno, and aborts in host builds to prevent silent double frees.

+++ include/sys/types.h
@@ -7,7 +7,7 @@ extern "C" {
 
 typedef long long int useconds_t;
 typedef long long int time_t;
-typedef long pid_t;
+typedef int pid_t;
 typedef int key_t;
 
 #ifdef __cplusplus
